pull common SQL functions to AbstractSQL
- refactor to eliminate duplicate code between Postgres, Mysql, &
  Sqlite implementations
