+++ Idno/Data/MySQL.php
@@ -11,41 +11,8 @@
 
         use Idno\Core\Idno;
 
-        class MySQL extends \Idno\Core\DataConcierge
+        class MySQL extends AbstractSQL
         {
-
-            private $dbname;
-            private $dbuser;
-            private $dbpass;
-            private $dbhost;
-            private $dbport;
-
-            function __construct($dbuser = null, $dbpass = null, $dbname = null, $dbhost = null, $dbport = null)
-            {
-                $this->dbuser = $dbuser;
-                $this->dbpass = $dbpass;
-                $this->dbname = $dbname;
-                $this->dbhost = $dbhost;
-                $this->dbport = $dbport;
-
-                if (empty($dbuser)) {
-                    $this->dbuser = \Idno\Core\Idno::site()->config()->dbuser;
-                }
-                if (empty($dbpass)) {
-                    $this->dbpass = \Idno\Core\Idno::site()->config()->dbpass;
-                }
-                if (empty($dbname)) {
-                    $this->dbname = \Idno\Core\Idno::site()->config()->dbname;
-                }
-                if (empty($dbhost)) {
-                    $this->dbhost = \Idno\Core\Idno::site()->config()->dbhost;
-                }
-                if (empty($dbport)) {
-                    $this->dbport = \Idno\Core\Idno::site()->config()->dbport;
-                }
-
-                parent::__construct();
-            }
 
             function init()
             {
@@ -137,27 +104,6 @@
             }
 
             /**
-             * Retrieve version information from the schema
-             * @return array|bool
-             */
-            function getVersions()
-            {
-                try {
-                    $client = $this->client;
-                    /* @var \PDO $client */
-                    $statement = $client->prepare("select * from `versions`");
-                    if ($statement->execute()) {
-                        return $statement->fetchAll(\PDO::FETCH_OBJ);
-                    }
-                } catch (\Exception $e) {
-                    //\Idno\Core\Idno::site()->logging()->error($e->getMessage());
-                    error_log($e->getMessage());
-                }
-
-                return false;
-            }
-
-            /**
              * Optimize tables - this can reduce overall database storage space and query time
              * @return bool
              */
@@ -172,25 +118,6 @@
                 }
 
                 return false;
-            }
-
-            /**
-             * Handle the session in MySQL
-             */
-            function handleSession()
-            {
-                if (version_compare(phpversion(), '5.3', '>')) {
-                    $sessionHandler = new \Symfony\Component\HttpFoundation\Session\Storage\Handler\PdoSessionHandler(\Idno\Core\Idno::site()->db()->getClient(),
-                        array(
-                            'db_table'    => 'session',
-                            'db_id_col'   => 'session_id',
-                            'db_data_col' => 'session_value',
-                            'db_time_col' => 'session_time',
-                        )
-                    );
-
-                    session_set_save_handler($sessionHandler, true);
-                }
             }
 
             /**
@@ -222,26 +149,6 @@
             }
 
             /**
-             * Saves a Known entity to the database, returning the _id
-             * field on success.
-             *
-             * @param Entity $object
-             */
-
-            function saveObject($object)
-            {
-                if ($object instanceof \Idno\Common\Entity) {
-                    if ($collection = $object->getCollection()) {
-                        $array = $object->saveToArray();
-
-                        return $this->saveRecord($collection, $array);
-                    }
-                }
-
-                return false;
-            }
-
-            /**
              * Saves a record to the specified database collection
              *
              * @param string $collection
@@ -269,8 +176,10 @@
                 if (empty($array['owner'])) {
                     $array['owner'] = '';
                 }
+
                 try {
                     $contents = json_encode($array);
+
                 } catch (\Exception $e) {
                     $contents = json_encode([]);
                     \Idno\Core\Idno::site()->logging()->error($e->getMessage());
@@ -375,27 +284,6 @@
             }
 
             /**
-             * Retrieves a Known entity object by its UUID, casting it to the
-             * correct class
-             *
-             * @param string $id
-             * @return \Idno\Common\Entity | false
-             */
-
-            function getObject($uuid)
-            {
-                if ($result = $this->getRecordByUUID($uuid)) {
-                    if ($object = $this->rowToEntity($result)) {
-                        if ($object->canRead()) {
-                            return $object;
-                        }
-                    }
-                }
-
-                return false;
-            }
-
-            /**
              * Retrieves a record from the database by its UUID
              *
              * @param string $id
@@ -466,80 +354,6 @@
                 }
 
                 return false;
-            }
-
-            /**
-             * Retrieve objects of a certain kind that we're allowed to see,
-             * (or excluding kinds that we don't want to see),
-             * in reverse chronological order
-             *
-             * @param string|array $subtypes String or array of subtypes we're allowed to see
-             * @param array $search Any extra search terms in array format (eg array('foo' => 'bar')) (default: empty)
-             * @param array $fields An array of fieldnames to return (leave empty for all; default: all)
-             * @param int $limit Maximum number of records to return (default: 10)
-             * @param int $offset Number of records to skip (default: 0)
-             * @param string $collection Collection to query; default: entities
-             * @param array $readGroups Which ACL groups should we check? (default: everything the user can see)
-             * @return array|false Array of elements or false, depending on success
-             */
-
-            function getObjects($subtypes = '', $search = array(), $fields = array(), $limit = 10, $offset = 0, $collection = 'entities', $readGroups = [])
-            {
-
-                // Initialize query parameters to be an empty array
-                $query_parameters = array();
-
-                // Ensure subtypes are recorded properly
-                // and remove subtypes that have an exclamation mark before them
-                // from consideration
-                if (!empty($subtypes)) {
-                    $not = array();
-                    if (!is_array($subtypes)) {
-                        $subtypes = array($subtypes);
-                    }
-                    foreach ($subtypes as $key => $subtype) {
-                        if (substr($subtype, 0, 1) == '!') {
-                            unset($subtypes[$key]);
-                            $not[] = substr($subtype, 1);
-                        }
-                    }
-                    if (!empty($subtypes)) {
-                        $query_parameters['entity_subtype']['$in'] = $subtypes;
-                    }
-                    if (!empty($not)) {
-                        $query_parameters['entity_subtype']['$not'] = $not;
-                    }
-                }
-
-                // Make sure we're only getting objects that we're allowed to see
-                if (empty($readGroups)) {
-                    $readGroups = \Idno\Core\Idno::site()->session()->getReadAccessGroupIDs();
-                }
-                $query_parameters['access'] = array('$in' => $readGroups);
-
-                // Join the rest of the search query elements to this search
-                $query_parameters = array_merge($query_parameters, $search);
-
-                // Prepare the fields array for searching, if required
-                if (!empty($fields) && is_array($fields)) {
-                    $fields = array_flip($fields);
-                    $fields = array_fill_keys($fields, true);
-                } else {
-                    $fields = array();
-                }
-
-                // Run the query
-                if ($results = $this->getRecords($fields, $query_parameters, $limit, $offset, $collection)) {
-                    $return = array();
-                    foreach ($results as $row) {
-                        $return[] = $this->rowToEntity($row);
-                    }
-
-                    return $return;
-                }
-
-                return array();
-
             }
 
             /**
@@ -642,38 +456,46 @@
                                 $variables[":value{$metadata_joins}"] = $value;
                             }
                         } else {
-                            if (!empty($value['$or'])) {
-                                $subwhere[] = "(" . $this->build_where_from_array($value['$or'], $variables, $metadata_joins, $non_md_variables, 'or', $collection) . ")";
-                            }
                             if (!empty($value['$not'])) {
                                 if (!empty($value['$not']['$in'])) {
-                                    $value['$not'] = array_merge($value['$not'], $value['$not']['$in']);
-                                    unset($value['$not']['$in']);
-                                }
-                                if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
-                                    $notstring = "`{$collection}`.`$key` not in(";
-                                    $i         = 0;
-                                    foreach ($value['$not'] as $val) {
-                                        if ($i > 0) $notstring .= ', ';
-                                        $notstring .= ":nonmdvalue{$non_md_variables}";
-                                        $variables[":nonmdvalue{$non_md_variables}"] = $val;
-                                        $non_md_variables++;
-                                        $i++;
+
+                                    if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
+                                        $notstring = "`{$collection}`.`$key` not in (";
+                                        $i         = 0;
+                                        foreach ($value['$not']['$in'] as $val) {
+                                            if ($i > 0) $notstring .= ', ';
+                                            $notstring .= ":nonmdvalue{$non_md_variables}";
+                                            $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                            $non_md_variables++;
+                                            $i++;
+                                        }
+                                        $notstring .= ")";
+                                    } else {
+                                        $metadata_joins++;
+                                        $notstring                           = "(md{$metadata_joins}.`name` = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` not in (";
+                                        $variables[":name{$metadata_joins}"] = $key;
+                                        $i                                   = 0;
+                                        foreach ($value['$not']['$in'] as $val) {
+                                            if ($i > 0) $notstring .= ', ';
+                                            $notstring .= ":nonmdvalue{$non_md_variables}";
+                                            $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                            $non_md_variables++;
+                                            $i++;
+                                        }
+                                        $notstring .= "))";
                                     }
-                                    $notstring .= ")";
                                 } else {
-                                    $metadata_joins++;
-                                    $notstring                           = "(md{$metadata_joins}.`name` = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` not in (";
-                                    $variables[":name{$metadata_joins}"] = $key;
-                                    $i                                   = 0;
-                                    foreach ($value['$not'] as $val) {
-                                        if ($i > 0) $notstring .= ', ';
-                                        $notstring .= ":nonmdvalue{$non_md_variables}";
-                                        $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                    if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
+                                        $notstring                                   = "`{$collection}`.`$key` != :nonmdvalue{$non_md_variables}";
+                                        $variables[":nonmdvalue{$non_md_variables}"] = $value['$not'];
                                         $non_md_variables++;
-                                        $i++;
+                                    } else {
+                                        $metadata_joins++;
+                                        $notstring = "(md{$metadata_joins}.`name`    = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` != :nonmdvalue{$non_md_variables})";
+                                        $variables[":name{$metadata_joins}"]         = $key;
+                                        $variables[":nonmdvalue{$non_md_variables}"] = $value['$not'];
+                                        $non_md_variables++;
                                     }
-                                    $notstring .= "))";
                                 }
                                 $subwhere[] = $notstring;
                             }
@@ -704,6 +526,9 @@
                                     $instring .= "))";
                                 }
                                 $subwhere[] = $instring;
+                            }
+                            if ($key == '$or') {
+                                $subwhere[] = "(" . $this->build_where_from_array($value, $variables, $metadata_joins, $non_md_variables, 'or', $collection) . ")";
                             }
                             if ($key == '$search') {
                                 if (!empty($value[0])) {
@@ -800,52 +625,6 @@
             }
 
             /**
-             * Count objects of a certain kind that we're allowed to see
-             *
-             * @param string|array $subtypes String or array of subtypes we're allowed to see
-             * @param array $search Any extra search terms in array format (eg array('foo' => 'bar')) (default: empty)
-             * @param string $collection Collection to query; default: entities
-             */
-            function countObjects($subtypes = '', $search = array(), $collection = 'entities')
-            {
-
-                // Initialize query parameters to be an empty array
-                $query_parameters = array();
-
-                // Ensure subtypes are recorded properly
-                // and remove subtypes that have an exclamation mark before them
-                // from consideration
-                if (!empty($subtypes)) {
-                    $not = array();
-                    if (!is_array($subtypes)) {
-                        $subtypes = array($subtypes);
-                    }
-                    foreach ($subtypes as $key => $subtype) {
-                        if (substr($subtype, 0, 1) == '!') {
-                            unset($subtypes[$key]);
-                            $not[] = substr($subtype, 1);
-                        }
-                    }
-                    if (!empty($subtypes)) {
-                        $query_parameters['entity_subtype']['$in'] = $subtypes;
-                    }
-                    if (!empty($not)) {
-                        $query_parameters['entity_subtype']['$not'] = $not;
-                    }
-                }
-
-                // Make sure we're only getting objects that we're allowed to see
-                $readGroups                 = \Idno\Core\Idno::site()->session()->getReadAccessGroupIDs();
-                $query_parameters['access'] = array('$in' => $readGroups);
-
-                // Join the rest of the search query elements to this search
-                $query_parameters = array_merge($query_parameters, $search);
-
-                return $this->countRecords($query_parameters, $collection);
-
-            }
-
-            /**
              * Count the number of records that match the given parameters
              * @param array $parameters
              * @param string $collection The collection to interrogate (default: 'entities')
@@ -889,19 +668,6 @@
             }
 
             /**
-             * Get database errors
-             * @return mixed
-             */
-            function getErrors()
-            {
-                if (!empty($this->client)) {
-                    return $this->client->errorInfo();
-                }
-
-                return false;
-            }
-
-            /**
              * Remove an entity from the database
              * @param string $id
              * @return true|false
@@ -932,36 +698,6 @@
                 return false;
             }
 
-            /**
-             * Retrieve the filesystem associated with the current db, suitable for saving
-             * and retrieving files
-             * @return bool
-             */
-            function getFilesystem()
-            {
-                // We're not returning a filesystem for MySQL
-                return false;
-            }
-
-            /**
-             * Given a text query, return an array suitable for adding into getFromX calls
-             * @param $query
-             * @return array
-             */
-            function createSearchArray($query)
-            {
-                return array('$search' => array($query));
-            }
-
-        }
-
-        /**
-         * Helper function that returns the current database object
-         * @return \Idno\Core\DataConcierge
-         */
-        function db()
-        {
-            return \Idno\Core\Idno::site()->db();
         }
 
     }
