+++ Idno/Data/Sqlite3.php
@@ -10,22 +10,8 @@
 
     namespace Idno\Data {
 
-        class Sqlite3 extends \Idno\Core\DataConcierge
+        class Sqlite3 extends AbstractSQL
         {
-
-            private $dbname;
-
-            function __construct($dbname = null)
-            {
-
-                $this->dbname = $dbname;
-
-                if (empty($dbname)) {
-                    $this->dbname = \Idno\Core\Idno::site()->config()->dbname;
-                }
-
-                parent::__construct();
-            }
 
             function init()
             {
@@ -110,53 +96,6 @@
                 }
 
                 return false;
-            }
-
-            /**
-             * Handle the session in Sqlite3
-             */
-            function handleSession()
-            {
-                if (version_compare(phpversion(), '5.3', '>')) {
-                    $sessionHandler = new \Symfony\Component\HttpFoundation\Session\Storage\Handler\PdoSessionHandler(\Idno\Core\Idno::site()->db()->getClient(),
-                        array(
-                            'db_table'    => 'session',
-                            'db_id_col'   => 'session_id',
-                            'db_data_col' => 'session_value',
-                            'db_time_col' => 'session_time',
-                        )
-                    );
-
-                    session_set_save_handler($sessionHandler, true);
-                }
-            }
-
-            /**
-             * Returns an instance of the database reference variable
-             * @return string;
-             */
-            function getDatabase()
-            {
-                return $this->database;
-            }
-
-            /**
-             * Returns an instance of the database client reference variable
-             * @return \PDO
-             */
-            function getClient()
-            {
-                return $this->client;
-            }
-
-            /**
-             * Sqlite3 doesn't need the ID to be processed.
-             * @param $id
-             * @return string
-             */
-            function processID($id)
-            {
-                return $id;
             }
 
             /**
@@ -349,80 +288,6 @@
             }
 
             /**
-             * Retrieve objects of a certain kind that we're allowed to see,
-             * (or excluding kinds that we don't want to see),
-             * in reverse chronological order
-             *
-             * @param string|array $subtypes String or array of subtypes we're allowed to see
-             * @param array $search Any extra search terms in array format (eg array('foo' => 'bar')) (default: empty)
-             * @param array $fields An array of fieldnames to return (leave empty for all; default: all)
-             * @param int $limit Maximum number of records to return (default: 10)
-             * @param int $offset Number of records to skip (default: 0)
-             * @param string $collection Collection to query; default: entities
-             * @param array $readGroups Which ACL groups should we check? (default: everything the user can see)
-             * @return array|false Array of elements or false, depending on success
-             */
-
-            function getObjects($subtypes = '', $search = array(), $fields = array(), $limit = 10, $offset = 0, $collection = 'entities', $readGroups = [])
-            {
-
-                // Initialize query parameters to be an empty array
-                $query_parameters = array();
-
-                // Ensure subtypes are recorded properly
-                // and remove subtypes that have an exclamation mark before them
-                // from consideration
-                if (!empty($subtypes)) {
-                    $not = array();
-                    if (!is_array($subtypes)) {
-                        $subtypes = array($subtypes);
-                    }
-                    foreach ($subtypes as $key => $subtype) {
-                        if (substr($subtype, 0, 1) == '!') {
-                            unset($subtypes[$key]);
-                            $not[] = substr($subtype, 1);
-                        }
-                    }
-                    if (!empty($subtypes)) {
-                        $query_parameters['entity_subtype']['$in'] = $subtypes;
-                    }
-                    if (!empty($not)) {
-                        $query_parameters['entity_subtype']['$not'] = $not;
-                    }
-                }
-
-                // Make sure we're only getting objects that we're allowed to see
-                if (empty($readGroups)) {
-                    $readGroups = \Idno\Core\Idno::site()->session()->getReadAccessGroupIDs();
-                }
-                $query_parameters['access'] = array('$in' => $readGroups);
-
-                // Join the rest of the search query elements to this search
-                $query_parameters = array_merge($query_parameters, $search);
-
-                // Prepare the fields array for searching, if required
-                if (!empty($fields) && is_array($fields)) {
-                    $fields = array_flip($fields);
-                    $fields = array_fill_keys($fields, true);
-                } else {
-                    $fields = array();
-                }
-
-                // Run the query
-                if ($results = $this->getRecords($fields, $query_parameters, $limit, $offset, $collection)) {
-                    $return = array();
-                    foreach ($results as $row) {
-                        $return[] = $this->rowToEntity($row);
-                    }
-
-                    return $return;
-                }
-
-                return array();
-
-            }
-
-            /**
              * Retrieves a set of records from the database with given parameters, in
              * reverse chronological order
              *
@@ -527,38 +392,47 @@
                                 $variables[":value{$metadata_joins}"] = $value;
                             }
                         } else {
-                            if (!empty($value['$or'])) {
-                                $subwhere[] = "(" . $this->build_where_from_array($value['$or'], $variables, $metadata_joins, $non_md_variables, 'or', $collection) . ")";
-                            }
                             if (!empty($value['$not'])) {
                                 if (!empty($value['$not']['$in'])) {
-                                    $value['$not'] = array_merge($value['$not'], $value['$not']['$in']);
-                                    unset($value['$not']['$in']);
+                                    if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
+                                        $notstring = "`{$collection}`.`$key` not in(";
+                                        $i         = 0;
+                                        foreach ($value['$not']['$in'] as $val) {
+                                            if ($i > 0) $notstring .= ', ';
+                                            $notstring .= ":nonmdvalue{$non_md_variables}";
+                                            $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                            $non_md_variables++;
+                                            $i++;
+                                        }
+                                        $notstring .= ")";
+                                    } else {
+                                        $metadata_joins++;
+                                        $notstring                           = "(md{$metadata_joins}.`name` = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` not in (";
+                                        $variables[":name{$metadata_joins}"] = $key;
+                                        $i                                   = 0;
+                                        foreach ($value['$not']['$in'] as $val) {
+                                            if ($i > 0) $notstring .= ', ';
+                                            $notstring .= ":nonmdvalue{$non_md_variables}";
+                                            $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                            $non_md_variables++;
+                                            $i++;
+                                        }
+                                        $notstring .= "))";
+                                    }
                                 }
-                                if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
-                                    $notstring = "`{$collection}`.`$key` not in(";
-                                    $i         = 0;
-                                    foreach ($value['$not'] as $val) {
-                                        if ($i > 0) $notstring .= ', ';
-                                        $notstring .= ":nonmdvalue{$non_md_variables}";
-                                        $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                // simple $not
+                                else {
+                                    if (in_array($key, array('uuid', '_id', 'entity_subtype', 'owner'))) {
+                                        $notstring                                   = "`{$collection}`.`$key` != :nonmdvalue{$non_md_variables}";
+                                        $variables[":nonmdvalue{$non_md_variables}"] = $value['$not'];
                                         $non_md_variables++;
-                                        $i++;
-                                    }
-                                    $notstring .= ")";
-                                } else {
-                                    $metadata_joins++;
-                                    $notstring                           = "(md{$metadata_joins}.`name` = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` not in (";
-                                    $variables[":name{$metadata_joins}"] = $key;
-                                    $i                                   = 0;
-                                    foreach ($value['$not'] as $val) {
-                                        if ($i > 0) $notstring .= ', ';
-                                        $notstring .= ":nonmdvalue{$non_md_variables}";
-                                        $variables[":nonmdvalue{$non_md_variables}"] = $val;
+                                    } else {
+                                        $metadata_joins++;
+                                        $notstring = "(md{$metadata_joins}.`name`    = :name{$metadata_joins} and md{$metadata_joins}.`collection` = '{$collection}' and md{$metadata_joins}.`value` != :nonmdvalue{$non_md_variables})";
+                                        $variables[":name{$metadata_joins}"]         = $key;
+                                        $variables[":nonmdvalue{$non_md_variables}"] = $value['$not'];
                                         $non_md_variables++;
-                                        $i++;
                                     }
-                                    $notstring .= "))";
                                 }
                                 $subwhere[] = $notstring;
                             }
@@ -590,6 +464,9 @@
                                 }
                                 $subwhere[] = $instring;
                             }
+                            if ($key == '$or') {
+                                $subwhere[] = "(" . $this->build_where_from_array($value, $variables, $metadata_joins, $non_md_variables, 'or', $collection) . ")";
+                            }
                             if ($key == '$search' && !empty($value)) {
                                 $val = $value[0]; // The search query is always in $value position [0] for now
 //                                if (strlen($val) > 5) {
@@ -609,52 +486,6 @@
                 }
 
                 return $where;
-            }
-
-            /**
-             * Count objects of a certain kind that we're allowed to see
-             *
-             * @param string|array $subtypes String or array of subtypes we're allowed to see
-             * @param array $search Any extra search terms in array format (eg array('foo' => 'bar')) (default: empty)
-             * @param string $collection Collection to query; default: entities
-             */
-            function countObjects($subtypes = '', $search = array(), $collection = 'entities')
-            {
-
-                // Initialize query parameters to be an empty array
-                $query_parameters = array();
-
-                // Ensure subtypes are recorded properly
-                // and remove subtypes that have an exclamation mark before them
-                // from consideration
-                if (!empty($subtypes)) {
-                    $not = array();
-                    if (!is_array($subtypes)) {
-                        $subtypes = array($subtypes);
-                    }
-                    foreach ($subtypes as $key => $subtype) {
-                        if (substr($subtype, 0, 1) == '!') {
-                            unset($subtypes[$key]);
-                            $not[] = substr($subtype, 1);
-                        }
-                    }
-                    if (!empty($subtypes)) {
-                        $query_parameters['entity_subtype']['$in'] = $subtypes;
-                    }
-                    if (!empty($not)) {
-                        $query_parameters['entity_subtype']['$not'] = $not;
-                    }
-                }
-
-                // Make sure we're only getting objects that we're allowed to see
-                $readGroups                 = \Idno\Core\Idno::site()->session()->getReadAccessGroupIDs();
-                $query_parameters['access'] = array('$in' => $readGroups);
-
-                // Join the rest of the search query elements to this search
-                $query_parameters = array_merge($query_parameters, $search);
-
-                return $this->countRecords($query_parameters, $collection);
-
             }
 
             /**
@@ -704,19 +535,6 @@
             }
 
             /**
-             * Get database errors
-             * @return mixed
-             */
-            function getErrors()
-            {
-                if (!empty($this->client)) {
-                    return $this->client->errorInfo();
-                }
-
-                return false;
-            }
-
-            /**
              * Remove an entity from the database
              * @param string $id
              * @return true|false
@@ -747,42 +565,10 @@
                 return false;
             }
 
-            /**
-             * Retrieve the filesystem associated with the current db, suitable for saving
-             * and retrieving files
-             * @return bool
-             */
-            function getFilesystem()
-            {
-                // We're not returning a filesystem for Sqlite3
-                return false;
-            }
-
-            /**
-             * Given a text query, return an array suitable for adding into getFromX calls
-             * @param $query
-             * @return array
-             */
-            function createSearchArray($query)
-            {
-                return array('$search' => array($query));
-            }
-
             public function exportRecords($collection = 'entities')
             {
-
-
                 return false; // TODO
             }
-        }
-
-        /**
-         * Helper function that returns the current database object
-         * @return \Idno\Core\DataConcierge
-         */
-        function db()
-        {
-            return \Idno\Core\Idno::site()->db();
         }
 
     }
